Fixes after the rebase from master
