+++ sysl2/naive/nestedGrammars_test.go
@@ -1,4 +1,4 @@
-package main
+package parser
 
 import (
 	"fmt"
